+++ src/main/java/insane96mcp/iguanatweaksreborn/module/movement/NoPillaring.java
@@ -7,7 +7,6 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import net.minecraft.core.BlockPos;
-import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.ai.targeting.TargetingConditions;
 import net.minecraft.world.entity.monster.Monster;
 import net.minecraft.world.entity.player.Player;
@@ -48,10 +47,10 @@ public class NoPillaring extends Feature {
 		BlockPos placedPos = event.getPos().relative(event.getFace());
 		Vec3 placedBlock = new Vec3(placedPos.getX() + 0.5d, placedPos.getY() + 0.5d, placedPos.getZ() + 0.5d);
 		double distance = placedBlock.distanceTo(playerEntity.position());
-		double allowedDistance = 1.36d;
-		if (playerEntity.hasEffect(MobEffects.JUMP))
+		//double allowedDistance = 1.36d;
+		//if (playerEntity.hasEffect(MobEffects.JUMP))
 			//noinspection ConstantConditions
-			allowedDistance *= 1 + ((playerEntity.getEffect(MobEffects.JUMP).getAmplifier() + 1) * 0.5);
+			//allowedDistance *= 1 + ((playerEntity.getEffect(MobEffects.JUMP).getAmplifier() + 1) * 0.5);
 		boolean isSolidBlock = true;
 		if (event.getItemStack().getItem() instanceof BlockItem) {
 			Block block = ((BlockItem) event.getItemStack().getItem()).getBlock();
@@ -60,7 +59,7 @@ public class NoPillaring extends Feature {
 				state = block.defaultBlockState();
 			isSolidBlock = state.blocksMotion();/*state.canOcclude();/*state.entityCanStandOn(event.getLevel(), event.getPos(), event.getEntity());*/
 		}
-		if (isSolidBlock && playerEntity.getViewXRot(1.0f) > 40f && !playerEntity.onGround() && event.getItemStack().getItem() instanceof BlockItem && distance <= allowedDistance && playerEntity.getY() > placedPos.getY()) {
+		if (isSolidBlock && playerEntity.getViewXRot(1.0f) > 40f && !playerEntity.onGround() && event.getItemStack().getItem() instanceof BlockItem && /*distance <= allowedDistance &&*/ playerEntity.getY() > placedPos.getY()) {
 			event.setCanceled(true);
 			event.setResult(Event.Result.DENY);
 		}